+++ backend/clinic-media/src/main/java/com/spring/clinicmedia/infrastructure/repositoryJpa/DoctorJpaRepository.java
@@ -12,7 +12,7 @@ import java.util.Optional;
 
 @Repository
 @EnableJpaRepositories
-public interface DoctorJpaRepository extends JpaRepository<Doctor, Integer> {
+public interface DoctorJpaRepository extends JpaRepository<Doctor, Long> {
 
 
     Optional<Doctor> findByRegistrationEmail(String email);
