+++ backend/clinic-media/src/main/java/com/spring/clinicmedia/infrastructure/repositoryJpa/LabJpaRepository.java
@@ -12,7 +12,7 @@ import java.util.Optional;
 
 @Repository
 @EnableJpaRepositories(basePackageClasses = Lab.class)
-public interface LabJpaRepository extends JpaRepository<Lab, Integer> {
+public interface LabJpaRepository extends JpaRepository<Lab, Long> {
 
 
     Optional<Lab> findLabByRegistrationEmail(String email);
