+++ backend/clinic-media/src/main/java/com/spring/clinicmedia/infrastructure/repositoryJpa/PatientJpaRepository.java
@@ -12,7 +12,7 @@ import java.util.Optional;
 
 @Repository
 @EnableJpaRepositories
-public interface PatientJpaRepository extends JpaRepository<Patient, Integer> {
+public interface PatientJpaRepository extends JpaRepository<Patient, Long> {
 
     Optional<Patient> findByRegistrationEmail(String email);
